requirements.txt
streamlit
openai
requests
beautifulsoup4
openpyxl
python-pptx
PyPDF2
lxml